adicionado parametro para produto, com um unico grupo
