+++ src/Estoque/saldos/campos/produto.c
@@ -31,11 +31,33 @@ int est_sld_prod_fun()
 		return 1;
 	}
 
-	find_subgrupos_restrict->posicao = 0;
-	find_subgrupos_restrict->entry = est_sld_grp_cod_entry;
-
-
-	gtk_widget_grab_focus(est_sld_grp_cod_entry);
+	if(!produtos.parametros.prod_varios_grupos){
+		sprintf(query,"select grupo from produtos where code = %s",est_sld_prod_cod_gchar);
+		if(!(estado = consultar(query))){
+			popup(NULL,"Erro ao consulta subgrupo");
+			return 1;
+		}
+		if(!(campo = mysql_fetch_row(estado))){
+			popup(NULL,"O produto inserido não possui grupo de produto");
+			return 1;
+		}
+		if(campo[0]){
+			est_sld_subgrp_cod_gchar = malloc(strlen(campo[0]));
+			strcpy(est_sld_subgrp_cod_gchar,campo[0]);
+			find_subgrupos_restrict->grupo = atoi(est_sld_subgrp_cod_gchar);
+		}else{
+			popup(NULL,"O grupo do produto inserido não pode ser encontrado");
+			return 1;
+		}
+		if(!calcula_saldo())
+			return 0;
+		else
+			return 1;
+	}else{
+		find_subgrupos_restrict->posicao = 0;
+		find_subgrupos_restrict->entry = est_sld_grp_cod_entry;
+		gtk_widget_grab_focus(est_sld_grp_cod_entry);
+	}
 
 	return 0;
 }
