+++ src/Estoque/saldos/campos/subgrupo.c
@@ -1,9 +1,9 @@
-int est_sald_subgrp_fun()
-{
-	char query[MAX_QUERY_LEN], valor[MAX_PRECO_LEN*2], **familia_char;
-	char *source,*dest;
+int est_sald_subgrp_fun(){
+	char query[MAX_QUERY_LEN];
 	MYSQL_RES *estado;
 	MYSQL_ROW campo;
+
+	char *source,*dest, **familia_char;
 	int grupo_len;
 
 	gtk_entry_set_text(GTK_ENTRY(est_sld_prod_entry),"");
@@ -47,45 +47,8 @@ int est_sald_subgrp_fun()
 	}
 
 	gtk_entry_set_text(GTK_ENTRY(est_sld_grp_nome_entry),dest);
-	gtk_entry_set_text(GTK_ENTRY(est_sld_prod_entry),"");
-
-	sprintf(query,"select SUM(entradas) - SUM(saidas),(select date_format(data_mov,'%cd/%cm/%cY') from movimento_estoque order by data_mov desc limit 1) from movimento_estoque where produto = %s and subgrupo = %s and estoque = %i",
-	37,37,37,
-	est_sld_prod_cod_gchar,
-	est_sld_subgrp_cod_gchar,
-	est_sld_prod_est_int);
-
-	if((estado = consultar(query))==NULL)
-		return 1;
-
-	if((campo = mysql_fetch_row(estado))==NULL)
-	{
-		popup(NULL,"O produto está sem saldo no estoque");
-		gtk_entry_set_text(GTK_ENTRY(est_sld_prod_entry),"");
-		gtk_entry_set_text(GTK_ENTRY(est_sld_data_entry),"");
-		return 1;
-	}
-	if(campo[0])
-	{
-		sprintf(valor,"%.3f",atof(campo[0]));
-		gtk_entry_set_text(GTK_ENTRY(est_sld_prod_entry),valor);
-		gtk_entry_set_text(GTK_ENTRY(est_sld_data_entry),campo[1]);
-	}
+	if(calcula_saldo())
+		return 0;
 	else
-	{
-		popup(NULL,"O produto nunca foi movimentado");
 		return 1;
-	}
-
-	sprintf(query,"select saldo_min from saldo_min_grupo where produto = %i and grupo = %i",atoi(est_sld_prod_cod_gchar),atoi(est_sld_subgrp_cod_gchar));
-	if((estado = consultar(query))==NULL)
-		return 1;
-
-	if((campo = mysql_fetch_row(estado))!=NULL)
-	{
-		sprintf(valor,"%.3f",atof(campo[0]));
-		gtk_entry_set_text(GTK_ENTRY(est_sld_min_entry),valor);
-	}
-
-	return 0;
 }
