+++ src/Parametros/ter_params.h
@@ -1,0 +1,24 @@
+static struct
+{
+	struct parametros
+	{
+		int doc;
+		int inscricao;
+		int tipodoc;
+		int bairro;
+		int endereco;
+		int cep;
+		int cidade;
+		int uf;
+		int tipo;
+		int grupo;
+		int observacoes;
+		int prazo;
+		int vlr_frete_pago;
+		int produto;
+		int preco;
+		int entrega;
+		int total;
+
+	}criticar;
+}terceiros;
