+++ src/etc/GerarDocs/Orcs/itens_por_grupo.h
@@ -1,0 +1,1 @@
+int gerar_orc_itens_por_grupos(FILE *file, struct _orc *orc);
