+++ src/etc/tools.c
@@ -549,78 +549,6 @@ int tasker(char *table)
 	return (task_num+1);
 }
 
-
-int configurar_parametros()
-{
-	int cont;
-	char *query;
-	query = malloc(MAX_QUERY_LEN);
-	MYSQL_RES *res;
-	MYSQL_ROW  row;
-
-  xmlDocPtr orc_xml = xmlReadFile(ORC_PARAMS, "UTF-8", XML_PARSE_RECOVER );
-	if(orc_xml){
-		xmlNodePtr root = xmlDocGetRootElement(orc_xml);
-		xmlNodePtr orc_prod_mov = getContentByTagName(root,"orc_prod_mov");
-		if(orc_prod_mov){
-			orcamentos.criticar.prod_movimento = atoi((const char *)orc_prod_mov->content);
-		}
-
-		xmlNodePtr orc_prod_saldo = getContentByTagName(root,"orc_prod_saldo");
-		if(orc_prod_saldo){
-			orcamentos.criticar.prod_saldo = atoi((const char *)orc_prod_saldo->content);
-		}
-
-		xmlNodePtr orc_prod_sld_lmt = getContentByTagName(root,"orc_prod_sld_lmt");
-		if(orc_prod_sld_lmt){
-			orcamentos.criticar.prod_saldo_limite = atoi((const char *)orc_prod_sld_lmt->content);
-		}
-
-		xmlNodePtr orc_ped_canc = getContentByTagName(root,"orc_ped_canc");
-		if(orc_ped_canc){
-			orcamentos.criticar.orc_ped_cancelado = atoi((const char *)orc_ped_canc->content);
-		}
-	}
-
-	sprintf(query,"select * from confs where code = %i",sessao_oper.code);
-
-	if((res = consultar(query))==NULL)
-	{
-		popup(NULL,"Erro ao receber dados de configuração");
-		return 1;
-	}
-
-	if((row = mysql_fetch_row(res))==NULL)
-	{
-		popup(NULL,"Sem dados para configurar o sistema");
-		return 1;
-	}
-
-	strcpy(navegadores.navegador_path1,row[1]);
-	strcpy(navegadores.navegador_path2,row[2]);
-	navegadores.navegador_pdr = atoi(row[3]);
-
-	strcpy(impressoras.imp_path1,row[4]);
-	strcpy(impressoras.imp_path2,row[5]);
-	strcpy(impressoras.imp_path3,row[6]);
-
-	carregar_navimps();
-
-	sprintf(query,"select * from orc_param");
-	if((res = consultar(query))==NULL)
-	{
-		popup(NULL,"Erro ao receber parametros de orçamentos");
-		return 1;
-	}
-
-	if((row = mysql_fetch_row(res))==NULL)
-	{
-		popup(NULL,"Sem dados para parametrizar orçamentos");
-		return 1;
-	}
-	orc_params.est_orc_padrao = atoi(row[0]);
-	return 0;
-}
 static char **tipo_ter_list;
 
 char *tipo_ter_num_to_str(int num){
